Bug Fixes and Input Behavior Enhancement
This commit addresses two main issues. Firstly, it fixes an error with the label position in the icon when 'informationAssociated' and 'disabled' are used, and an error when pressing the button in 'informationAssociated'. The styles have been fixed and unnecessary 'onStructureBlur/focus' have been removed. Secondly, it introduces a change to the 'useInput' hook to prevent value changes when the mouse wheel is used on a number input, avoiding unexpected behavior for the user.