Add custom content feature to FunctionalitiesModule and ListOptions components
This commit introduces a new feature that allows the FunctionalitiesModule and ListOptions components to display custom content. This is achieved by adding a `content` property to these components and their corresponding types. The `content` property is of type `React.ReactNode`, allowing for flexible rendering of custom content.

Updates have been made to the corresponding test and story files to reflect this new feature. A new story, `FunctionalitiesModuleWithContent`, has been added to demonstrate the usage of the new feature.

These changes enhance the flexibility of the FunctionalitiesModule and ListOptions components, allowing them to be used in a wider range of contexts within the application.