Enhance Handling of Input Values in useInput Hook
This commit improves the handling of input values in the useInput hook, particularly in the context of formatting, validation, and control.

Changes include:

The truncateValue function in useInput.ts now takes into account the locale or formatNumber.locale in addition to limitedValue and maxDecimals. This enhances the truncation of the value with maximum decimals.
The handleChangeInternal function in useInput.ts has been refactored to use a new local variable eventValue instead of directly manipulating event.target.value. This improves code readability and maintainability.
The condition to assign valueControlled to event.target.value in useInput.ts has been modified to exclude InputTypeType.NUMBER. This improves the handling of input values for different types.
These changes enhance the functionality and readability of the useInput hook.