Refactor device height handling with new useDeviceHeight hook
This commit introduces a significant refactoring of the application's handling of device height. A new hook, useDeviceHeight, has been introduced and is now used across multiple components to set the height. The previous viewportHeight mixin has been removed and replaced with a CSS variable --100svh that is set by the new hook. This refactoring aims to enhance the application's responsiveness to changes in device height. Components such as ActionBottomSheet and Drawer have been updated to use the new hook and CSS variable. Tests for the new useDeviceHeight hook have also been added.