Update onIconClick prop in components
This commit fixes the onIconClick prop so it now works as expected. However, it's important to note that this prop will be deprecated in the next major release. Instead, users should transition to using activeIcon onClick and disabledIcon onClick for handling icon click events. This change is part of our ongoing efforts to improve the API and usability of our components.